refactor vue setup
