--- site/vue/src/components/sqlToolboxToolBar.vue
@@ -1,1 +1,0 @@
-<template>temp</template>
