+++ site/vue/src/main.ts
@@ -2,15 +2,26 @@ import { createApp } from 'vue';
 import Unknown from './components/Unknown.vue';
 
 const exports = {
-    async render(target: string | Element, component: string, args: Record<string, unknown> = {}) {
+    async render(
+        target: string | Element,
+        type: 'component' | 'page',
+        name: string,
+        args: Record<string, unknown> = {},
+    ) {
         const app = await (async () => {
             try {
-                // eslint-disable-next-line no-unsanitized/method
-                return createApp((await import(`./components/${component}.vue`) as { default: Parameters<typeof createApp>[0] }).default, args);
+                // https://vite.dev/guide/features.html#glob-import
+                const modules = import.meta.glob(['./components/**/*.vue', './components/*.vue', './pages/*.vue']);
+                const path = `./${type}s/${name}.vue`;
+                if (!(path in modules)) {
+                    throw new Error(`Module ${path} not found`);
+                }
+                const mod = await modules[path]() as { default: Parameters<typeof createApp>[0] };
+                return createApp(mod.default, args);
             }
             catch (e) {
-                console.error(`Could not find vue component ${component}:`, e);
-                return createApp(Unknown, { component });
+                console.error(`Could not find vue ${type} ${name}:`, e);
+                return createApp(Unknown, { type, name });
             }
         })();
