+++ site/vue/tsconfig.app.json
@@ -1,6 +1,6 @@
 {
   "extends": "@vue/tsconfig/tsconfig.dom.json",
-  "include": ["env.d.ts", "src/**/*", "src/**/*.vue", "../ts/**/*"],
+  "include": ["env.d.ts", "src/**/*", "src/**/*.vue", "src/ts/*.ts","../ts/**/*"],
   "exclude": ["src/**/__tests__/*"],
   "compilerOptions": {
     "composite": true,
